Address @kainino0x feedback
